+++ src/kesiswaan/attendance/entities/attendance-summary.entity.ts
@@ -16,19 +16,19 @@ export class AttendanceSummary {
   @Column({ type: 'varchar', length: 7 })
   tahun_bulan: string; // "2025-01"
 
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'int', default: 0 })
   total_hadir: number;
 
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'int', default: 0 })
   total_sakit: number;
 
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'int', default: 0 })
   total_izin: number;
 
-  @Column({ type: 'tinyint', default: 0 })
+  @Column({ type: 'int', default: 0 })
   total_alpa: number;
 
-  @Column({ type: 'tinyint' })
+  @Column({ type: 'int' })
   total_days_expected: number;
 
   @Column({ type: 'decimal', precision: 5, scale: 2 })
